System.Wordexp

Description

wordexp (and wordfree) Haskell wrapper

Wrapper

wordexp :: Flags -> String -> IO (Either WordexpError [String])

wordexp wrapper

Allows to specify desired flags, return expanded strings or encountered error if any

Flags

data Flags

wordexp flags

Not every flag is supported since some of them do not make much sense in Haskell anyway

nosubst :: Flags

Disable command substitution in patterns, treat them as errors

errors :: Flags

Do not hide shell error messages in devnull, print them right away

noundef :: Flags

Do not accept undefined shell variables, treat them as errors

Errors

data WordexpError

Possible wordexp errors

Constructors

OutOfSpace

IllegalCharacterOccurence

UndefinedShellVariable

CommandSubstitution

ShellSyntaxError